ViNTT update BE

+++ Project-c1120g1/src/main/java/com/c1120g1/adweb/entity/Post.java
@@ -56,8 +56,8 @@ public class Post {
 
     @Min(0)
     @Max(2000000000)
-    @Column(name = "price", columnDefinition = "DOUBLE")
-    private Long price;
+    @Column(name = "price", columnDefinition = "INT")
+    private Integer price;
 
     @NotBlank
     @Column(name = "description", columnDefinition = "TEXT")
